backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Spark/Delta (for sync CLI)
pyspark==3.5.2
delta-spark==3.2.0
